Meet minimum requirements for the app
User can enter a search term, the app gets thumbnail images from
the YouTube API, and they display in a decent-looking, responsive layout
on the web page. Clicking on a thumbnail opens the video in a new tab.

To do:
Get the videos to play in a lightbox.
Add a button for each thumbnail to get more from that channel.
Add a buttons to get more results (using the previous and next
page links from the JSON).